correction of point y

+++ src/data-sturctures/Point.h
@@ -7,7 +7,7 @@ class Point
 {
     public:
     uint16_t x;
-    uint8_t y;
+    uint16_t y;
     Point(uint16_t = 0, uint16_t = 0);
 };
